LayoutTests/media/volume-sleep-disable.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <title>volume-active-audio-session</title>
    <script src="video-test.js"></script>
    <script src="media-file.js"></script>
    <script>
    async function runTest() {
        findMediaElement();
        run('video.volume = 0');
        run('video.src = findMediaFile("video", "content/test-25fps")');
        await run('video.play()');
        testExpected('internals.elementIsBlockingDisplaySleep(video)', false);
        run('video.volume = 1');
        await testExpectedEventually('internals.elementIsBlockingDisplaySleep(video)', true, '==', 1000);
    }

    window.addEventListener('load', event => {
        runTest().then(endTest).catch(failTest);
    })
    </script>
</head>
<body>
    <video controls playsinline></video>
</body>
</html>
